paddlex pipeline not working for table
